requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-3-jet==1.0.8
django-4-jet==1.0.8
django-admin-interface==0.19.1
django-clearcache==1.2.1
django-colorfield==0.7.0
django-flat-json-widget==0.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-jet-reboot==1.3.1
django-json-widget==1.1.1
django-prettyjson==0.4.1
djongo==1.3.6
dnspython==2.2.1
docopt==0.6.2
et-xmlfile==1.1.0
future==0.18.2
idna==3.3
MarkupPy==1.14
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.2
Pillow==9.1.1
pipreqs==0.4.11
pymongo==3.12.3
PyPDF2==2.0.0
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
standardjson==0.3.1
tablib==3.2.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.10
uuid==1.30
winkerberos==0.9.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
